RH values float32

+++ const/__init__.py
@@ -1,4 +1,5 @@
-rh_type = {f'rh{i}': 'float64' for i in range(100)}
+#%%
+rh_type = {f'rh{i}': 'float32' for i in range(100)}
 dtypes = {
     'system:index': 'object',
     'beam': 'uint16',
@@ -38,3 +39,4 @@ dtypes = {
     '.geo': 'object'
 }
 dtypes.update(rh_type)
+# %%
